src/requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
spacy
pypdf
pytesseract
pillow
uvicorn
